requirements.txt
streamlit>=1.32.0
boto3>=1.34.0
python-dotenv>=1.0.0
omnara>=0.1.0
